Move Ray and Plane into core.
