+++ amethyst_core/src/lib.rs
@@ -46,6 +46,7 @@ pub use self::{
 pub mod bundle;
 pub mod deferred_dispatcher_operation;
 pub mod frame_limiter;
+pub mod geometry;
 pub mod timing;
 pub mod transform;
